# openapi.yaml
openapi: 3.0.3
info:
  title: IssueFlow API
  description: |
    Ticket management backend platform — TDP 2026 Home Assignment.

    Manages users, projects, tickets (issues), comments, attachments, ticket
    dependencies, audit logs, CSV import/export, soft-deletes, @mentions,
    auto-assignment to the least-loaded developer, and scheduled priority
    escalation for overdue tickets.

    Authentication: stateless JWT bearer token obtained from `POST /auth/login`.
  version: 1.0.0
  contact:
    name: IssueFlow Team
    email: rm-TDPIsrael@intl.att.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://issueflow.example.com
    description: Production (placeholder)
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserResponse' }
    post:
      tags: [Users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get a user by id
      parameters: [ { $ref: '#/components/parameters/UserId' } ]
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Users]
      summary: Delete a user
      parameters: [ { $ref: '#/components/parameters/UserId' } ]
      responses:
        '200': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/update/{userId}:
    patch:
      tags: [Users]
      summary: Update a user's full name and/or role
      parameters: [ { $ref: '#/components/parameters/UserId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200': { description: Updated }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT issued by `POST /auth/login`. Send as `Authorization: Bearer <token>`."
  parameters:
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: Validation failure or malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Concurrent update conflict (optimistic lock)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, DEVELOPER]
...
    ApiError:
      type: object
      required: [timestamp, status, error, path]
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-05-20T10:15:30Z'
        status:
          type: integer
          example: 400
        error:
          type: string
          example: 'Bad Request'
        message:
          type: string
          example: 'username must not be blank'
        path:
          type: string
          example: '/users'
        traceId:
          type: string
          description: OpenTelemetry trace id for correlation
          example: '4bf92f3577b34da6a3ce929d0e0e4736'
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string, example: 'email' }
              message: { type: string, example: 'must be a well-formed email address' }

    # ---------- Auth ----------
...
    UserResponse:
      type: object
      required: [id, username, email, fullName, role]
      properties:
        id:       { type: integer, format: int64, example: 1 }
        username: { type: string, example: jdoe }
        email:    { type: string, format: email, example: jdoe@example.com }
        fullName: { type: string, example: John Doe }
        role:     { $ref: '#/components/schemas/Role' }
    UserCreateRequest:
      type: object
      required: [username, email, fullName, role]
      properties:
        username: { type: string, minLength: 3, maxLength: 50, pattern: '^[A-Za-z0-9_]+$', example: jdoe }
        email:    { type: string, format: email, example: jdoe@example.com }
        fullName: { type: string, minLength: 1, maxLength: 120, example: John Doe }
        role:     { $ref: '#/components/schemas/Role' }
        password: { type: string, format: password, minLength: 8, example: s3cretP@ss }
    UserUpdateRequest:
      type: object
      properties:
        fullName: { type: string, minLength: 1, maxLength: 120, example: Jane Doe }
        role:     { $ref: '#/components/schemas/Role' }
...